Change message after a successful installation
Related issue: https://github.com/BlackCorsair/install-jdk-on-steam-deck/issues/5

For some time, the `bashrc` file in steamos has changed
and now whenever `source ~/.bashrc`
is run, it gives some prompt errors.

Until this is sorted out or fixed, the
message after a successful installation
has been changed so it suggests
opening a new terminal or a
new shell in order for the changes
to be loaded.